THRIFT-244. cpp: Fix TJSONProtocol::writeMessageBegin
uint8_t is a character type, so lexical_cast doesn't convert it
to an integer representation.


git-svn-id: https://svn.apache.org/repos/asf/incubator/thrift/trunk@734862 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/cpp/src/protocol/TJSONProtocol.cpp
@@ -32,7 +32,7 @@ static const uint8_t kJSONEscapeChar = 'u';
 
 static const std::string kJSONEscapePrefix("\\u00");
 
-static const uint8_t kThriftVersion1 = 1;
+static const uint32_t kThriftVersion1 = 1;
 
 static const std::string kThriftNan("NaN");
 static const std::string kThriftInfinity("Infinity");
